+++ src/redball/scenes/main/AbstractScene.java
@@ -3,5 +3,5 @@ package redball.scenes.main;
 public abstract class AbstractScene {
     public void start() {}
 
-    public abstract void update(double deltaTime);
+    public abstract void update(float deltaTime);
 }
